+++ macrobond_data_api/util/__init__.py
@@ -3,3 +3,5 @@ Utility methods for Macrobond Data API.
 """
 
 from .save_credential_to_keyring import save_credential_to_keyring
+
+from .transfer_performance_test import transfer_performance_test
